sway: run commands without waiting for Xwayland
Xwayland is lazy now, there is no need to wait at all
